improve project structure
