+++ .github/workflows/main.yml
@@ -25,14 +25,6 @@ jobs:
           cargo clippy --all-targets --features=tor -- -D warnings
           cargo clippy --all-targets --no-default-features -- -D warnings
           cargo clippy --all-targets --no-default-features --features=tor -- -D warnings
-          cd test_utils
-          cargo clippy -- -D warnings
-          cargo clippy --features=tokio -- -D warnings
-          cd ..
-          cd integration-tests
-          cargo clippy -- -D warnings
-          cargo clippy --no-default-features -- -D warnings
-          cd ..
       - name: Install 3proxy
         run: |
           cd $HOME
@@ -62,6 +54,5 @@ jobs:
           cargo build --verbose --all --no-default-features --features=tor
           cargo test --lib --verbose --no-default-features --features=tor
       - name: Integration tests
-        working-directory: ./integration-tests
         run: |
           env PATH=$HOME/3proxy-0.8.13/src:$PATH tests/integration_tests.sh
