+++ Cargo.toml
@@ -17,11 +17,11 @@ travis-ci = { repository = "sticnarf/tokio-socks" }
 
 [[example]]
 name = "chainproxy"
-required-features = []
+required-features = ["tokio"]
 
 [[example]]
 name = "socket"
-required-features = ["tor"]
+required-features = ["tor", "tokio"]
 
 [[example]]
 name = "tor"
@@ -42,7 +42,8 @@ thiserror = "1.0"
 [dev-dependencies]
 futures-executor = "0.3"
 tokio = { version = "1.0", features = ["io-util", "rt-multi-thread", "net"] }
-async-std = { version = "1.12" }
+once_cell = { version = "1.2" }
+tokio-util = { version = "0.7", features = ["compat"] }
 
 [features]
 default = ["tokio"]
