--- integration-tests/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "integration-tests"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-tokio = { version = "1.0", features = [
-    "io-util",
-    "rt-multi-thread",
-    "net",
-], optional = true }
-test_utils = { path = "../test_utils" }
-tokio-socks = { path = "..", default-features = false }
-
-[features]
-default = ["use_tokio"]
-use_tokio = ["tokio", "test_utils/tokio", "tokio-socks/tokio"]
