+++ src/tcp/socks4/futures_impl.rs
@@ -1,0 +1,28 @@
+//! This module contains futures-io-specfic implementations.
+use super::*;
+
+impl<T> AsyncRead for Socks4Stream<T>
+where
+    T: AsyncRead + Unpin,
+{
+    fn poll_read(mut self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &mut [u8]) -> Poll<io::Result<usize>> {
+        AsyncRead::poll_read(Pin::new(&mut self.socket), cx, buf)
+    }
+}
+
+impl<T> AsyncWrite for Socks4Stream<T>
+where
+    T: AsyncWrite + Unpin,
+{
+    fn poll_write(mut self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &[u8]) -> Poll<io::Result<usize>> {
+        AsyncWrite::poll_write(Pin::new(&mut self.socket), cx, buf)
+    }
+
+    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
+        AsyncWrite::poll_flush(Pin::new(&mut self.socket), cx)
+    }
+
+    fn poll_close(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
+        AsyncWrite::poll_close(Pin::new(&mut self.socket), cx)
+    }
+}
