+++ src/tcp/socks4/mod.rs
@@ -1,4 +1,9 @@
-use crate::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
+#[cfg(not(feature = "tokio"))]
+mod futures_impl;
+#[cfg(feature = "tokio")]
+mod tokio_impl;
+
+use crate::io::{read_exact, AsyncRead, AsyncWrite, AsyncWriteExt};
 use crate::{Error, IntoTargetAddr, Result, TargetAddr};
 use futures_util::stream::{self, Fuse, Stream, StreamExt};
 use std::{
@@ -9,8 +14,6 @@ use std::{
     pin::Pin,
     task::{Context, Poll},
 };
-#[cfg(feature = "tokio")]
-use tokio::net::TcpStream;
 
 #[repr(u8)]
 #[derive(Clone, Copy)]
@@ -39,70 +42,6 @@ impl<S> Deref for Socks4Stream<S> {
 impl<S> DerefMut for Socks4Stream<S> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.socket
-    }
-}
-
-#[cfg(feature = "tokio")]
-use crate::ToProxyAddrs;
-
-#[cfg(feature = "tokio")]
-impl Socks4Stream<TcpStream> {
-    /// Connects to a target server through a SOCKS4 proxy given the proxy
-    /// address.
-    ///
-    /// # Error
-    ///
-    /// It propagates the error that occurs in the conversion from `T` to
-    /// `TargetAddr`.
-    pub async fn connect<'t, P, T>(proxy: P, target: T) -> Result<Socks4Stream<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        Self::execute_command(proxy, target, None, CommandV4::Connect).await
-    }
-
-    /// Connects to a target server through a SOCKS4 proxy using given username,
-    /// password and the address of the proxy.
-    ///
-    /// # Error
-    ///
-    /// It propagates the error that occurs in the conversion from `T` to
-    /// `TargetAddr`.
-    pub async fn connect_with_userid<'a, 't, P, T>(
-        proxy: P,
-        target: T,
-        user_id: &'a str,
-    ) -> Result<Socks4Stream<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        Self::execute_command(proxy, target, Some(user_id), CommandV4::Connect).await
-    }
-
-    async fn execute_command<'a, 't, P, T>(
-        proxy: P,
-        target: T,
-        user_id: Option<&'a str>,
-        command: CommandV4,
-    ) -> Result<Socks4Stream<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        Self::validate_userid(user_id)?;
-
-        let sock = Socks4Connector::new(
-            user_id,
-            command,
-            proxy.to_proxy_addrs().fuse(),
-            target.into_target_addr()?,
-        )
-        .execute()
-        .await?;
-
-        Ok(sock)
     }
 }
 
@@ -220,9 +159,9 @@ where
 
     #[cfg(feature = "tokio")]
     /// Connect to the proxy server, authenticate and issue the SOCKS command
-    pub async fn execute(&mut self) -> Result<Socks4Stream<TcpStream>> {
+    pub async fn execute(&mut self) -> Result<Socks4Stream<tokio::net::TcpStream>> {
         let next_addr = self.proxy.select_next_some().await?;
-        let tcp = TcpStream::connect(next_addr)
+        let tcp = tokio::net::TcpStream::connect(next_addr)
             .await
             .map_err(|_| Error::ProxyServerUnreachable)?;
 
@@ -309,15 +248,7 @@ where
         */
 
         self.prepare_recv_reply();
-        #[cfg(feature = "tokio")]
-        {
-            self.ptr += tcp.read_exact(&mut self.buf[self.ptr..self.len]).await?;
-        }
-        #[cfg(not(feature = "tokio"))]
-        {
-            tcp.read_exact(&mut self.buf[self.ptr..self.len]).await?;
-            self.ptr = self.len;
-        }
+        self.ptr += read_exact(tcp, &mut self.buf[self.ptr..self.len]).await?;
 
         if self.buf[0] != 0 {
             return Err(Error::InvalidResponseVersion);
@@ -347,69 +278,6 @@ where
 /// address.
 pub struct Socks4Listener<S> {
     inner: Socks4Stream<S>,
-}
-
-#[cfg(feature = "tokio")]
-impl Socks4Listener<TcpStream> {
-    /// Initiates a BIND request to the specified proxy.
-    ///
-    /// The proxy will filter incoming connections based on the value of
-    /// `target`.
-    ///
-    /// # Error
-    ///
-    /// It propagates the error that occurs in the conversion from `T` to
-    /// `TargetAddr`.
-    pub async fn bind<'t, P, T>(proxy: P, target: T) -> Result<Socks4Listener<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        Self::bind_to_target(None, proxy, target).await
-    }
-
-    /// Initiates a BIND request to the specified proxy using given username
-    /// and password.
-    ///
-    /// The proxy will filter incoming connections based on the value of
-    /// `target`.
-    ///
-    /// # Error
-    ///
-    /// It propagates the error that occurs in the conversion from `T` to
-    /// `TargetAddr`.
-    pub async fn bind_with_userid<'a, 't, P, T>(
-        proxy: P,
-        target: T,
-        user_id: &'a str,
-    ) -> Result<Socks4Listener<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        Self::bind_to_target(Some(user_id), proxy, target).await
-    }
-
-    async fn bind_to_target<'a, 't, P, T>(
-        user_id: Option<&'a str>,
-        proxy: P,
-        target: T,
-    ) -> Result<Socks4Listener<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        let socket = Socks4Connector::new(
-            user_id,
-            CommandV4::Bind,
-            proxy.to_proxy_addrs().fuse(),
-            target.into_target_addr()?,
-        )
-        .execute()
-        .await?;
-
-        Ok(Socks4Listener { inner: socket })
-    }
 }
 
 impl<S> Socks4Listener<S>
@@ -504,66 +372,5 @@ where
             socket: self.inner.socket,
             target,
         })
-    }
-}
-
-#[cfg(not(feature = "tokio"))]
-mod futures_io {
-    use super::*;
-    impl<T> AsyncRead for Socks4Stream<T>
-    where
-        T: AsyncRead + Unpin,
-    {
-        fn poll_read(mut self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &mut [u8]) -> Poll<io::Result<usize>> {
-            AsyncRead::poll_read(Pin::new(&mut self.socket), cx, buf)
-        }
-    }
-
-    impl<T> AsyncWrite for Socks4Stream<T>
-    where
-        T: AsyncWrite + Unpin,
-    {
-        fn poll_write(mut self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &[u8]) -> Poll<io::Result<usize>> {
-            AsyncWrite::poll_write(Pin::new(&mut self.socket), cx, buf)
-        }
-
-        fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            AsyncWrite::poll_flush(Pin::new(&mut self.socket), cx)
-        }
-
-        fn poll_close(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            AsyncWrite::poll_close(Pin::new(&mut self.socket), cx)
-        }
-    }
-}
-
-#[cfg(feature = "tokio")]
-mod tokio_io {
-    use super::*;
-    use tokio::io::ReadBuf;
-    impl<T> AsyncRead for Socks4Stream<T>
-    where
-        T: AsyncRead + Unpin,
-    {
-        fn poll_read(mut self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &mut ReadBuf<'_>) -> Poll<io::Result<()>> {
-            AsyncRead::poll_read(Pin::new(&mut self.socket), cx, buf)
-        }
-    }
-
-    impl<T> AsyncWrite for Socks4Stream<T>
-    where
-        T: AsyncWrite + Unpin,
-    {
-        fn poll_write(mut self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &[u8]) -> Poll<io::Result<usize>> {
-            AsyncWrite::poll_write(Pin::new(&mut self.socket), cx, buf)
-        }
-
-        fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            AsyncWrite::poll_flush(Pin::new(&mut self.socket), cx)
-        }
-
-        fn poll_shutdown(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            AsyncWrite::poll_shutdown(Pin::new(&mut self.socket), cx)
-        }
     }
 }
