+++ src/tcp/socks5/mod.rs
@@ -1,4 +1,9 @@
-use crate::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
+#[cfg(not(feature = "tokio"))]
+mod futures_impl;
+#[cfg(feature = "tokio")]
+mod tokio_impl;
+
+use crate::io::{read_exact, AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
 use crate::{Authentication, Error, IntoTargetAddr, Result, TargetAddr};
 use futures_util::stream::{self, Fuse, Stream, StreamExt};
 use std::{
@@ -9,8 +14,6 @@ use std::{
     pin::Pin,
     task::{Context, Poll},
 };
-#[cfg(feature = "tokio")]
-use tokio::net::TcpStream;
 
 #[repr(u8)]
 #[derive(Clone, Copy)]
@@ -45,96 +48,6 @@ impl<S> Deref for Socks5Stream<S> {
 impl<S> DerefMut for Socks5Stream<S> {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.socket
-    }
-}
-
-#[cfg(feature = "tokio")]
-impl Socks5Stream<TcpStream> {
-    /// Connects to a target server through a SOCKS5 proxy given the proxy
-    /// address.
-    ///
-    /// # Error
-    ///
-    /// It propagates the error that occurs in the conversion from `T` to
-    /// `TargetAddr`.
-    pub async fn connect<'t, P, T>(proxy: P, target: T) -> Result<Socks5Stream<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        Self::execute_command(proxy, target, Authentication::None, Command::Connect).await
-    }
-
-    /// Connects to a target server through a SOCKS5 proxy using given username,
-    /// password and the address of the proxy.
-    ///
-    /// # Error
-    ///
-    /// It propagates the error that occurs in the conversion from `T` to
-    /// `TargetAddr`.
-    pub async fn connect_with_password<'a, 't, P, T>(
-        proxy: P,
-        target: T,
-        username: &'a str,
-        password: &'a str,
-    ) -> Result<Socks5Stream<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        Self::execute_command(
-            proxy,
-            target,
-            Authentication::Password { username, password },
-            Command::Connect,
-        )
-        .await
-    }
-
-    #[cfg(feature = "tor")]
-    /// Resolve the domain name to an ip using special Tor Resolve command, by
-    /// connecting to a Tor compatible proxy given it's address.
-    pub async fn tor_resolve<'t, P, T>(proxy: P, target: T) -> Result<TargetAddr<'static>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        let sock = Self::execute_command(proxy, target, Authentication::None, Command::TorResolve).await?;
-
-        Ok(sock.target_addr().to_owned())
-    }
-
-    #[cfg(feature = "tor")]
-    /// Perform a reverse DNS query on the given ip using special Tor Resolve
-    /// PTR command, by connecting to a Tor compatible proxy given it's
-    /// address.
-    pub async fn tor_resolve_ptr<'t, P, T>(proxy: P, target: T) -> Result<TargetAddr<'static>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        let sock = Self::execute_command(proxy, target, Authentication::None, Command::TorResolvePtr).await?;
-
-        Ok(sock.target_addr().to_owned())
-    }
-
-    async fn execute_command<'a, 't, P, T>(
-        proxy: P,
-        target: T,
-        auth: Authentication<'a>,
-        command: Command,
-    ) -> Result<Socks5Stream<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        Self::validate_auth(&auth)?;
-
-        let sock = SocksConnector::new(auth, command, proxy.to_proxy_addrs().fuse(), target.into_target_addr()?)
-            .execute()
-            .await?;
-
-        Ok(sock)
     }
 }
 
@@ -288,9 +201,9 @@ where
 
     #[cfg(feature = "tokio")]
     /// Connect to the proxy server, authenticate and issue the SOCKS command
-    pub async fn execute(&mut self) -> Result<Socks5Stream<TcpStream>> {
+    pub async fn execute(&mut self) -> Result<Socks5Stream<tokio::net::TcpStream>> {
         let next_addr = self.proxy.select_next_some().await?;
-        let tcp = TcpStream::connect(next_addr)
+        let tcp = tokio::net::TcpStream::connect(next_addr)
             .await
             .map_err(|_| Error::ProxyServerUnreachable)?;
 
@@ -439,15 +352,7 @@ where
 
     async fn receive_reply<T: AsyncRead + AsyncWrite + Unpin>(&mut self, tcp: &mut T) -> Result<TargetAddr<'static>> {
         self.prepare_recv_reply();
-        #[cfg(feature = "tokio")]
-        {
-            self.ptr += tcp.read_exact(&mut self.buf[self.ptr..self.len]).await?;
-        }
-        #[cfg(not(feature = "tokio"))]
-        {
-            tcp.read_exact(&mut self.buf[self.ptr..self.len]).await?;
-            self.ptr = self.len;
-        }
+        self.ptr += read_exact(tcp, &mut self.buf[self.ptr..self.len]).await?;
         if self.buf[0] != 0x05 {
             return Err(Error::InvalidResponseVersion);
         }
@@ -480,29 +385,13 @@ where
             // Domain
             0x03 => {
                 self.len = 5;
-                #[cfg(feature = "tokio")]
-                {
-                    self.ptr += tcp.read_exact(&mut self.buf[self.ptr..self.len]).await?;
-                }
-                #[cfg(not(feature = "tokio"))]
-                {
-                    tcp.read_exact(&mut self.buf[self.ptr..self.len]).await?;
-                    self.ptr = self.len;
-                }
+                self.ptr += read_exact(tcp, &mut self.buf[self.ptr..self.len]).await?;
                 self.len += self.buf[4] as usize + 2;
             },
             _ => Err(Error::UnknownAddressType)?,
         }
 
-        #[cfg(feature = "tokio")]
-        {
-            self.ptr += tcp.read_exact(&mut self.buf[self.ptr..self.len]).await?;
-        }
-        #[cfg(not(feature = "tokio"))]
-        {
-            tcp.read_exact(&mut self.buf[self.ptr..self.len]).await?;
-            self.ptr = self.len;
-        }
+        self.ptr += read_exact(tcp, &mut self.buf[self.ptr..self.len]).await?;
         let target: TargetAddr<'static> = match self.buf[3] {
             // IPv4
             0x01 => {
@@ -543,73 +432,6 @@ where
 /// address.
 pub struct Socks5Listener<S> {
     inner: Socks5Stream<S>,
-}
-
-#[cfg(feature = "tokio")]
-use crate::ToProxyAddrs;
-
-#[cfg(feature = "tokio")]
-impl Socks5Listener<TcpStream> {
-    /// Initiates a BIND request to the specified proxy.
-    ///
-    /// The proxy will filter incoming connections based on the value of
-    /// `target`.
-    ///
-    /// # Error
-    ///
-    /// It propagates the error that occurs in the conversion from `T` to
-    /// `TargetAddr`.
-    pub async fn bind<'t, P, T>(proxy: P, target: T) -> Result<Socks5Listener<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        Self::bind_with_auth(Authentication::None, proxy, target).await
-    }
-
-    /// Initiates a BIND request to the specified proxy using given username
-    /// and password.
-    ///
-    /// The proxy will filter incoming connections based on the value of
-    /// `target`.
-    ///
-    /// # Error
-    ///
-    /// It propagates the error that occurs in the conversion from `T` to
-    /// `TargetAddr`.
-    pub async fn bind_with_password<'a, 't, P, T>(
-        proxy: P,
-        target: T,
-        username: &'a str,
-        password: &'a str,
-    ) -> Result<Socks5Listener<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        Self::bind_with_auth(Authentication::Password { username, password }, proxy, target).await
-    }
-
-    async fn bind_with_auth<'t, P, T>(
-        auth: Authentication<'_>,
-        proxy: P,
-        target: T,
-    ) -> Result<Socks5Listener<TcpStream>>
-    where
-        P: ToProxyAddrs,
-        T: IntoTargetAddr<'t>,
-    {
-        let socket = SocksConnector::new(
-            auth,
-            Command::Bind,
-            proxy.to_proxy_addrs().fuse(),
-            target.into_target_addr()?,
-        )
-        .execute()
-        .await?;
-
-        Ok(Socks5Listener { inner: socket })
-    }
 }
 
 impl<S> Socks5Listener<S>
@@ -700,66 +522,5 @@ where
             socket: self.inner.socket,
             target,
         })
-    }
-}
-
-#[cfg(not(feature = "tokio"))]
-mod futures_io {
-    use super::*;
-    impl<T> AsyncRead for Socks5Stream<T>
-    where
-        T: AsyncRead + Unpin,
-    {
-        fn poll_read(mut self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &mut [u8]) -> Poll<io::Result<usize>> {
-            AsyncRead::poll_read(Pin::new(&mut self.socket), cx, buf)
-        }
-    }
-
-    impl<T> AsyncWrite for Socks5Stream<T>
-    where
-        T: AsyncWrite + Unpin,
-    {
-        fn poll_write(mut self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &[u8]) -> Poll<io::Result<usize>> {
-            AsyncWrite::poll_write(Pin::new(&mut self.socket), cx, buf)
-        }
-
-        fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            AsyncWrite::poll_flush(Pin::new(&mut self.socket), cx)
-        }
-
-        fn poll_close(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            AsyncWrite::poll_close(Pin::new(&mut self.socket), cx)
-        }
-    }
-}
-
-#[cfg(feature = "tokio")]
-mod tokio_io {
-    use super::*;
-    use tokio::io::ReadBuf;
-    impl<T> AsyncRead for Socks5Stream<T>
-    where
-        T: AsyncRead + Unpin,
-    {
-        fn poll_read(mut self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &mut ReadBuf<'_>) -> Poll<io::Result<()>> {
-            AsyncRead::poll_read(Pin::new(&mut self.socket), cx, buf)
-        }
-    }
-
-    impl<T> AsyncWrite for Socks5Stream<T>
-    where
-        T: AsyncWrite + Unpin,
-    {
-        fn poll_write(mut self: Pin<&mut Self>, cx: &mut Context<'_>, buf: &[u8]) -> Poll<io::Result<usize>> {
-            AsyncWrite::poll_write(Pin::new(&mut self.socket), cx, buf)
-        }
-
-        fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            AsyncWrite::poll_flush(Pin::new(&mut self.socket), cx)
-        }
-
-        fn poll_shutdown(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-            AsyncWrite::poll_shutdown(Pin::new(&mut self.socket), cx)
-        }
     }
 }
