--- test_utils/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "test_utils"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-futures-util = { version = "0.3", default-features = false, features = [
-    "io",
-    "std",
-] }
-tokio = { version = "1.0", features = ["io-util", "rt-multi-thread", "net"] }
-once_cell = { version = "1.2" }
-tokio-socks = { path = "../", default-features = false }
-tokio-util = { version = "0.7", features = ["compat"] }
-
-[features]
-tokio = ["tokio-socks/tokio"]
